(PC-20310)[API] refactor: use subcategories_v2 in the api
