+++ api/src/pcapi/core/bookings/models.py
@@ -32,7 +32,7 @@ from sqlalchemy.sql.elements import Label
 from pcapi.core.bookings import exceptions
 from pcapi.core.bookings.constants import BOOKINGS_AUTO_EXPIRY_DELAY
 from pcapi.core.bookings.constants import BOOKS_BOOKINGS_AUTO_EXPIRY_DELAY
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 import pcapi.core.finance.models as finance_models
 from pcapi.core.offers import models as offers_models
 from pcapi.models import Base
