+++ api/src/pcapi/core/finance/validation.py
@@ -2,7 +2,7 @@ import datetime
 
 import pytz
 
-from pcapi.core.categories.subcategories import ALL_SUBCATEGORIES_DICT
+from pcapi.core.categories.subcategories_v2 import ALL_SUBCATEGORIES_DICT
 import pcapi.utils.date as date_utils
 
 from . import exceptions
