+++ api/src/pcapi/core/mails/transactional/bookings/booking_soon_to_be_expired_to_beneficiary.py
@@ -4,7 +4,7 @@ from typing import Tuple
 from pcapi.core import mails
 from pcapi.core.bookings import constants as booking_constants
 from pcapi.core.bookings.models import Booking
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 from pcapi.core.mails import models
 from pcapi.core.mails.transactional.sendinblue_template_ids import TransactionalEmail
 from pcapi.core.users.models import User
