+++ api/src/pcapi/core/mails/transactional/pro/first_venue_approved_offer_to_pro.py
@@ -4,7 +4,7 @@ from sqlalchemy.orm import joinedload
 
 from pcapi.core import mails
 from pcapi.core.bookings import constants as booking_constants
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 from pcapi.core.mails import models
 from pcapi.core.mails.transactional.sendinblue_template_ids import TransactionalEmail
 from pcapi.core.offerers.models import Venue
