+++ api/src/pcapi/core/offers/factories.py
@@ -5,7 +5,7 @@ import uuid
 
 import factory
 
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 import pcapi.core.offerers.factories as offerers_factories
 import pcapi.core.offers.models as offers_models
 from pcapi.core.testing import BaseFactory
