+++ api/src/pcapi/sandboxes/scripts/creators/allocine/__init__.py
@@ -1,6 +1,6 @@
 from random import randint
 
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 import pcapi.core.offerers.factories as offerers_factories
 import pcapi.core.offers.factories as offers_factories
 import pcapi.core.providers.factories as providers_factories
