+++ api/src/pcapi/sandboxes/scripts/creators/big/create_big_offerer.py
@@ -2,7 +2,7 @@ import datetime
 import logging
 
 from pcapi.core.bookings.factories import BookingFactory
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 from pcapi.core.offerers.factories import UserOffererFactory
 from pcapi.core.offerers.factories import VenueFactory
 from pcapi.core.offerers.factories import VirtualVenueFactory
