+++ api/src/pcapi/sandboxes/scripts/creators/data/create_data_event_products.py
@@ -1,6 +1,6 @@
 import logging
 
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 import pcapi.core.offers.factories as offers_factories
 import pcapi.core.offers.models as offers_models
 from pcapi.domain.music_types import music_types
