+++ api/src/pcapi/sandboxes/scripts/creators/data/create_data_thing_products.py
@@ -1,7 +1,7 @@
 import logging
 import random
 
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 import pcapi.core.offers.factories as offers_factories
 import pcapi.core.offers.models as offers_models
 from pcapi.core.providers.titelive_gtl import GTLS
