+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_industrial_invoices.py
@@ -2,7 +2,7 @@ from datetime import datetime
 import logging
 
 import pcapi.core.bookings.factories as bookings_factories
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 from pcapi.core.finance import api as finance_api
 from pcapi.core.finance import factories as finance_factories
 from pcapi.core.finance import models as finance_models
