+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_offer_with_thousand_stocks.py
@@ -1,7 +1,7 @@
 import datetime
 import logging
 
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 from pcapi.core.offerers.factories import UserOffererFactory
 from pcapi.core.offerers.factories import VenueFactory
 from pcapi.core.offerers.factories import VirtualVenueFactory
