+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_offerer_providers_for_apis.py
@@ -5,7 +5,7 @@ import typing
 
 from pcapi import settings
 from pcapi.core.bookings import factories as bookings_factories
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 from pcapi.core.educational import factories as educational_factories
 from pcapi.core.external_bookings import factories as external_bookings_factories
 from pcapi.core.offerers import factories as offerers_factories
