+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_offers_with_ean.py
@@ -1,6 +1,6 @@
 import logging
 
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 from pcapi.core.criteria import factories as criteria_factories
 from pcapi.core.offerers import factories as offerers_factories
 from pcapi.core.offers import factories as offers_factories
