+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_offers_with_price_categories.py
@@ -1,6 +1,6 @@
 import logging
 
-from pcapi.core.categories import subcategories
+from pcapi.core.categories import subcategories_v2 as subcategories
 from pcapi.core.offerers.factories import UserOffererFactory
 from pcapi.core.offerers.factories import VenueFactory
 from pcapi.core.offerers.factories import VirtualVenueFactory
